Add files via upload
Created a Java dice game called Pig.  It's two players and the first player to reach 50 points wins.  The rules for the games are as follows. The goal of the game is to be the first to get fifty points or more. Players take turns rolling a six-sided die. If a one is rolled then the player gets no points and it becomes the other player's turn. If it is two through six then the player can do two things. Either roll again or hold to gain all of those points. 

Suggestions are welcomed.